book_parser/requirements.txt
pdf2image==1.17.0
pillow==12.0.0
PyPDF2==3.0.1
